+++ src/main/java/com/buuz135/thaumicjei/ThaumicJEI.java
@@ -8,7 +8,7 @@ import org.apache.logging.log4j.Logger;
         modid = ThaumicJEI.MOD_ID,
         name = ThaumicJEI.MOD_NAME,
         version = ThaumicJEI.VERSION,
-        dependencies = "required-after:jei@[1.12.2-4.9.1.169,);required-after:thaumcraft@[6.1.BETA10,);",
+        dependencies = "required-after:jei@[1.12.2-4.10.0.198,);required-after:thaumcraft@[6.1.BETA10,);",
         clientSideOnly = true
 )
 public class ThaumicJEI {
@@ -28,4 +28,8 @@ public class ThaumicJEI {
     public void onPre(FMLPreInitializationEvent event) {
         LOGGER = event.getModLog();
     }
+
+    //TODO Show a shapeless icon in the recipes that are shapeless
+    //TODO Fix gui screen
+    //TODO register subtypes for vis crystals and phials
 }